+++ SCrawler.PluginProvider/Interfaces/IPropertyProvider.vb
@@ -1,0 +1,13 @@
+﻿' Copyright (C) 2023  Andy https://github.com/AAndyProgram
+' This program is free software: you can redistribute it and/or modify
+' it under the terms of the GNU General Public License as published by
+' the Free Software Foundation, either version 3 of the License, or
+' (at your option) any later version.
+'
+' This program is distributed in the hope that it will be useful,
+' but WITHOUT ANY WARRANTY
+Namespace Plugin
+    Public Interface IPropertyProvider : Inherits IFormatProvider
+        Property PropertyName As String
+    End Interface
+End Namespace