+++ SCrawler/API/Twitter/Declarations.vb
@@ -12,6 +12,7 @@ Imports PersonalUtilities.Functions.RegularExpressions
 Namespace API.Twitter
     Friend Module Declarations
         Friend Const TwitterSite As String = "Twitter"
+        Friend Const TwitterSiteKey As String = "AndyProgram_Twitter"
         Friend ReadOnly DateProvider As ADateTime = GetDateProvider()
         Friend ReadOnly VideoNode As NodeParams() = {New NodeParams("video_info", True, True, True, True, 10)}
         Friend ReadOnly VideoSizeRegEx As RParams = RParams.DMS("\d+x(\d+)", 1, EDP.ReturnValue)